GrossSlimes and A Fast Skelly Breed to Gimmel
The two new mobs to spawn are slimes and skellies. However, these skeletons are fast, equipped with sharp poison II bones, and are extremely weak. They are like glass knives that cut deep but break easily. Also, they only spawn if the light level is 4 or below, so you would have to be in moonlight or sheltered from light to see them spawn.

As for the slimes, they spawn rarely and give you debuffs if they hit you. Stickiness (slowing you down with Slowness II) is assured, and poison from the gross slime is possible next.

Both mobs are placed into Gimmel's plains. Still, perhaps slimes aren't supposed to be introduced yet, but players have wanted some slimes for quite a while.