DarkJesus Collab to Further Buff Fire Watch
We replaced the diamond chestplate with leather, and the skelly is all leather armor now to reflect the lore of an ancient forest watch that is annoyed with birch forest destruction. The skelly now carries a powerful blaze rod that has knockback 8, sharpness 2. In addition, a hit by this item now induces blindness and wither effects.

Also, it's important to note that this skelly is wearing a black glass head. We are not sure if that line of code to add a glass head will work, but if it does not, then get rid of the glass head. If it works, then we have skellies with black glass heads.