DJ Collab to add Iron Golem and special witches to Bet
Like the previous pull, formatting of the name of the mobs may be interesting for testing the code. Also, the iron golem in this code only has a chance of dropping a fire aspect II sword, and it should not drop any iron. 

Also, there is a witch spawn set over the empty and ruined NSR. It's just a cobble encased town with grief all inside after a great war runined all of its newfriends. They are invisible, slow, and drop newfriend souls (netherwart lore) when they die. This should test if it's possible for witches to spawn in an area defined by a circle.

As for the iron golem, it is tough, and it may be set to create special, valuable ingredients in the future that require many players to kill it. I reccomend not hitting it if you want to live.