Dark Jesus Collab, Adding Gimmel Mobs
Changed Amthrax to anthrax in config spelling.

Added two new mobs for the plains biome. Also, added ghost witches to Aquila's location. They are supposed to spawn in town at night, they are slow, and they are invisible. They drop ten netherwarts that are newfriend souls when they die (the lore has no specific purpose for anything useful).

The new mobs are a dirty spider and a dark archer. THe first is a cave spider with some serious speed II, and debuffs like blindness and posion against players. They drop some good loot for how hard they are to kill. As for the dark archer, he drops sticky pistons, but he has regen III, and punch IV with fire I on the bow. The debuffs are blindness and then night vision for the player, and confusion.

The spawn rate in the desert should also buff up, with skellies remaining about the same in that area and other mobs like the solar creeper blowing more stuff up. It has seemed there aren't many mobs at all in the desert.