Glass helmet eliminated, Feather boots returned
Mobs can't wear foreign blocks like black glass, so the code is changed to let the skelly wear the fire prot leather for a helmet. Also, the boots have feather falling again to make this thing harder to die from falling.

Further, a small dropchance of 2% remains on all items that this skelly wears, but the special blaze rod will not drop. (Eliminating that itemheld dropchance option should still make it impossible for the skelly to drop that blaze rod.) Those armor sets could be a novel item to walk around in, and should take a lot of time to grind for. 

Also, the skelly should not be able to pick up items anyway, so I got rid of the option that set Pickupitems to false since it is redundant.