DarkArcher drops one sticky piston, not two now
DJ suggests that this monster needs to drop some valuable good like sticky pistons. The Dark Archer is less powerful than the iron golem, we think, so it will only drop one sticky piston. We will test if it is possible to kill these things, and the feedback will help later.